Move option utils into the options directory
Relocate src/plugin/src/utils/option-utils.js to src/plugin/src/options/option-utils.js so the option normalization helpers sit with the rest of the options code instead of a standalone utils folder. Update all imports to follow the new location.
